requirements.txt
accelerate==1.2.0
einops==0.8.1
hydra-core==1.3.2
librosa==0.10.2.post1
matplotlib==3.10.7
nemo_toolkit==2.1.0rc0
neptune==1.14.0
numpy==2.2.6
omegaconf==2.3.0
phonemizer==3.3.0
safetensors==0.6.2
soundfile==0.12.1
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
webrtcvad==2.0.10
seaborn==0.13.2
lightning==2.5.5
lhotse==1.31.1
sentencepiece==0.2.1
jiwer==4.0.0
pyannote.core==6.0.1
webdataset==1.0.2
datasets==4.2.0
editdistance==0.8.1
pyannote.metrics==4.0.0
IPython==8.37.0
wandb==0.22.2